move ordering to model package
